+++ transformers/examples/sentence_retrieval/run_sentence_retrieval_synthetic.py
@@ -205,6 +205,13 @@ class DataTrainingArguments:
             "help": "When used with one-to-one mapping, also changes the [CLS] and [SEP] token. Does not change the PAD token."
         },
     )
+    # Permutation
+    permute_words: bool = field(
+        default=False,
+        metadata={
+            "help": "Permute the words of the sentence randomly. Different permutation for each sentence."
+        },
+    )    
     # Word modification or syntax modifications
     # (If word modification, then monolingual corpus, if syntax modification, then parallel corpus)
     bilingual: bool = field(